backend/requirements.txt
langchain>=0.1.5
langchain-community>=0.0.35
# Thêm lại các thư viện Google API
google-generativeai>=0.7.1
google-api-core>=2.15.0
google-auth>=2.26.1
proto-plus>=1.23.0
protobuf>=4.25.1
cohere>=4.47
python-dotenv>=1.0.0
chromadb>=0.4.22
faiss-cpu>=1.7.4
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.6.3
python-multipart>=0.0.9
numpy>=1.24.0
passlib[bcrypt]>=1.7.4
langsmith>=0.1.0
requests>=2.31.0
# New dependencies for auth and progress tracking
flask>=2.0.1
flask-cors>=3.0.10
sqlalchemy>=2.0.0
werkzeug>=2.0.1
colorama>=0.4.4
flask-swagger-ui==4.11.1
openai==1.3.3
tiktoken==0.5.1
pymysql==1.1.0